- Refactored `year` field to `academicYear` across `User` model, `UserMapper`, and `UserServiceImpl` as it is a reserved keyword for h2 and other databases. - Added H2 database dependency and updated development configuration for H2 support for streamlined development. - Minor typo fix in POM comments. - Change "ServiceImpls" package to "implementations" as to fit package naming convention
