+++ src/main/java/com/tuconnect/dorm_connect/service/implementations/ChatServiceImpl.java
@@ -1,4 +1,4 @@
-package com.tuconnect.dorm_connect.service.ServiceImpl;
+package com.tuconnect.dorm_connect.service.implementations;
 
 import com.tuconnect.dorm_connect.dto.Chat.ChatDTO;
 import com.tuconnect.dorm_connect.dto.Messages.MessageDTO;
