+++ src/main/java/com/tuconnect/dorm_connect/service/implementations/UserServiceImpl.java
@@ -1,4 +1,4 @@
-package com.tuconnect.dorm_connect.service.ServiceImpl;
+package com.tuconnect.dorm_connect.service.implementations;
 
 import com.tuconnect.dorm_connect.dto.User.UserDTO;
 import com.tuconnect.dorm_connect.mapper.UserMapper;
@@ -42,7 +42,7 @@ public class UserServiceImpl implements UserService {
         user.setPassword(dto.password());
         user.setProfileImageUrl(dto.profileImageUrl());
         user.setMajor(dto.major());
-        user.setYear(dto.year());
+        user.setAcademicYear(dto.year());
         user.setRole(Roles.Users);
 
         User savedUser = userRepository.save(user);
@@ -61,7 +61,7 @@ public class UserServiceImpl implements UserService {
         existingUser.setPassword(updatedUserDTO.password());
         existingUser.setProfileImageUrl(updatedUserDTO.profileImageUrl());
         existingUser.setMajor(updatedUserDTO.major());
-        existingUser.setYear(updatedUserDTO.year());
+        existingUser.setAcademicYear(updatedUserDTO.year());
 
 
         User savedUser = userRepository.save(existingUser);
