+++ src/main/resources/application.properties
@@ -2,13 +2,24 @@ spring.application.name=TU Connect
 spring.profiles.active=dev
 
 # PostgreSQL Database Configuration for application-dev.properties
+# UNCOMMENT THESE FOR DEMOS
 #spring.datasource.url=jdbc:postgresql://localhost:5432/<DataBase>
 #spring.datasource.username=<User>
 #spring.datasource.password=<Password>
+#spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+
+# H2 Database Configuration
+# ONLY INTENDED FOR USAGE DURING DEVELOPMENT PURPOSES, COMMENT OUT IF USING POSTGRESQL
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2-console
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
 
 spring.jpa.hibernate.ddl-auto=create
 spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 
 # Swagger
 springdoc.api-docs.path=/api-docs
